build: :construction_worker: set publishing to npm build pipeline
