+++ src/export.ts
@@ -1,0 +1,3 @@
+import { tailwindcssClay } from "./main"
+
+export default tailwindcssClay
